+++ inst/doc/available-palettes.R
@@ -1,111 +1,111 @@
-## ---- include = FALSE---------------------------------------------------------
+## ----include = FALSE----------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  install.packages("PrettyCols")
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  remotes::install_github("nrennie/PrettyCols")
 
 ## -----------------------------------------------------------------------------
 library(PrettyCols)
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Blues")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Purples")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Tangerines")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Greens")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Pinks")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Teals")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Yellows")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Reds")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("PurpleGreens")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("PinkGreens")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("TangerineBlues")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("PurpleGreens")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("PurplePinks")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("TealGreens")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("PurpleYellows")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("RedBlues")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Bold")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Light")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Dark")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Neon")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Summer")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Autumn")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Winter")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Rainbow")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Beach")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Fun")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Sea")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Bright")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Relax")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Lucent")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Lively")
 
-## ---- out.width = "60%", fig.align="center"-----------------------------------
+## ----out.width = "60%", fig.align="center"------------------------------------
 prettycols("Joyful")