+++ NEWS.md
@@ -1,6 +1,23 @@
+## PrettyCols 1.1.0 2024_06_22
+
+### Breaking changes
+
+* Argument `name` in `scale_*_*()` functions is now used to define the title of the legend, to make this consistent with other {ggplot2} `scale` functions. If you have previously used e.g. `scale_fill_manual(name = "Bright")` to say you want to use the `"Bright"` palette, this will result in an error. Please use `scale_fill_manual(palette = "Bright")` instead.
+* Argument `legend_title` in `scale_*_*()` functions is deprecated. Please use `name` instead. For now, `legend_title` still works. 
+
+### Non-breaking changes
+
+* Move Python implementation to separate repository
+* Add `scale_*_pretty_div()` diverging scale functions
+* Changes examples to `mtcars` data
+* Remove `scale_name` from `discrete_scale()` calls
+* Add {lifecycle} to Imports
+* Removed {purrr} dependency
+* Add 8 new palettes
+
 ## PrettyCols 1.0.1 2023_01_27
 
-* Add new palettes: `Beach`, `Fun`
+* Add 12 new palettes
 * Allow multiple selection of palettes
 * Filter by colourblind friendly palettes
 * Add `legend_title` argument to scale functions